requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
sqlmodel==0.0.21
pydantic==2.9.2
pydantic-settings==2.4.0
Pillow==10.4.0
python-multipart==0.0.9
